backend/requirements.txt
fastapi
uvicorn

chromadb

sentence-transformers
torch
numpy

google-generativeai

requests
beautifulsoup4
lxml

python-dotenv

aiohttp